+++ CHANGELOG.md
@@ -3,7 +3,7 @@ title: "CHANGELOG"
 Version: 1.9.1
 Last Backup: 2026-03-02
 template-status: origin_dev
-modified: 2026-03-03
+modified: 2026-04-05
 ---
 ⬆️:: [[🏡Home]]
 [[TODO]] - [[BACKLOG]] - [[RELEASE NOTES]]
@@ -11,6 +11,10 @@ modified: 2026-03-03
 *Last updated: `= date(now)`*
 ---
 > Vault modifications and changes log.
+---
+# 05/04/26
+- Added [[obsidian-ooda-agent.js]]
+
 ---
 # 01/04/26
 ## Inbox Processing